Move auto-generated reducers to separate files
