+++ src/sentry/static/sentry/app/redux/reducers/event.jsx
@@ -1,0 +1,1 @@
+export const EVENTS = 'EVENTS';
