+++ src/sentry/static/sentry/app/redux/reducers/workBatchDetailsEntry.jsx
@@ -1,0 +1,14 @@
+import {resource} from 'app/redux/reducers/shared';
+
+export const WORK_BATCH_DETAILS = 'WORK_BATCH_DETAILS';
+
+const sharedInitialState = {
+  ...resource.initialState,
+};
+
+const workBatchDetailsEntry = resource.createReducer(
+  WORK_BATCH_DETAILS,
+  sharedInitialState
+);
+
+export default workBatchDetailsEntry;
